Tweaked delta export
